// src/node/manifest.d.ts
export interface WebAccessibleResourceByMatch {
  matches: string[]
  resources: string[]
  use_dynamic_url?: boolean
}

export interface WebAccessibleResourceById {
  extension_ids: string[]
  resources: string[]
  use_dynamic_url?: boolean
}